feat: validando usuarios autorizados

+++ back-end/src/use-cases/task/destroy-task .ts
@@ -1,18 +1,24 @@
 import { TasksRepository } from '@/repositories/tasks'
 import { ResourceNotFoundError } from '../errors/resource-not-found'
+import { OperationNotPermitedError } from '../errors/operation-not-permited-error'
 
 interface DestroyTaskUseCaseRequest {
   taskId: string
+  userId: string
 }
 
 export class DestroyTaskUseCase {
   constructor(private tasksRepository: TasksRepository) {}
 
-  async execute({ taskId }: DestroyTaskUseCaseRequest) {
+  async execute({ taskId, userId }: DestroyTaskUseCaseRequest) {
     const task = await this.tasksRepository.findById(taskId)
 
     if (!task) {
       throw new ResourceNotFoundError()
+    }
+
+    if (task.userId !== userId) {
+      throw new OperationNotPermitedError()
     }
 
     await this.tasksRepository.destroy(taskId)
